Use cpm package types
